[nexus] convert external API to trait
At long last, here's the last OpenAPI document converted into a trait: the
Nexus external API. This is a fairly large change, but with all the
prerequisites in place the conversion is overall pretty mechanical. I did need
to move some types around.

In a subsequent PR I'll:

1. Remove the old instructions.
2. Add a `.git-blame-ignore-revs` entry for each of the API commits, to make
   blame better.

Depends on (TODO).