[openapi-manager] richer extra validation
With the Nexus external API, the validator generates a `nexus_tags.txt` file
that must be kept track of.  Instead of the validation function simply erroring
out if the file is different, it is a better experience for users if it records
what it expects the file to be, and then the OpenAPI manager simply treats it
as an extra file similar to the document itself.

With this pattern, the check and generate functions can both work on the extra
file just like they work on the document.

In order for the there to be a richer protocol for validation, the interface
needs to be split into its own crate. This way, the API crates can depend on
this minimal interface, and the OpenAPI manager itself can depend on the API
crates.

Pull Request: https://github.com/oxidecomputer/omicron/pull/6370